Feature/et 24 write readme (#6)
* Initial setup

* Remove resource actions

* Adding main Onfido class / Switching calls to be instance instead of static

* ET-13: Making changes based on PR feedback

* Initial commit for ET-13

* ET-13: Adding back in unkownApiUrl config property

* ET-13: Adding .travis.yml

* ET-13: Fixing location of .travis.yml

* ET-13: Fixing failing tests

* ET-13: Added all actions for the applicant class

* ET-13: Fixing failing test

* ET-15: Added basic upload and download functions / Two actions added for documents as tests for the new functions

* ET-16: Initial commit for ET-16

* ET-15: Making chnages based on PR feedback

* ET-15: Fixing failing test

* ET-16: All Manager classes finished

* ET-16: All test classes complete

* ET-15: Fixing wildcard imports

* ET-16: Added Address, SdkToken, and Webhook endpoints

* ET-17: Added webhook verification

* ET-24: ReadMe complete

* ET-17: Making changes based on PR feedback

* ET-17: Fixing test

* ET-17: Making changes per PR feedback

* ET-24: adding WIP warning to readme

* ET-24: Updating readme based on PR feedback

* ET-24: Fixing minor typo

* ET-24: Removing missed note

* ET-24: Fixing typo

* ET-24: Removing resource section Per Ciaran's suggestion

* ET-24: removing .java and adding error handling example
