+++ onfido-java/src/main/java/com/onfido/api/WebhookEventObject.java
@@ -1,0 +1,56 @@
+package com.onfido.api;
+
+import com.squareup.moshi.Json;
+
+import java.util.Objects;
+
+public class WebhookEventObject {
+
+    @Json(name = "id") private final String id;
+    @Json(name = "status") private final String status;
+    @Json(name = "href") private final String href;
+    @Json(name = "completed_at_iso8601") private final String completedAtIso8601;
+
+    protected WebhookEventObject(String id, String status, String href, String completedAtIso8601) {
+        this.id = id;
+        this.status = status;
+        this.href = href;
+        this.completedAtIso8601 = completedAtIso8601;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public String getStatus() {
+        return status;
+    }
+
+    public String getHref() {
+        return href;
+    }
+
+    public String getCompletedAtIso8601() {
+        return completedAtIso8601;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        WebhookEventObject object = (WebhookEventObject) o;
+        return getId().equals(object.getId()) &&
+                getStatus().equals(object.getStatus()) &&
+                getHref().equals(object.getHref()) &&
+                getCompletedAtIso8601().equals(object.getCompletedAtIso8601());
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(id, href, status, completedAtIso8601);
+    }
+}
