+++ onfido-java/src/main/java/com/onfido/api/WebhookPayload.java
@@ -1,0 +1,16 @@
+package com.onfido.api;
+
+import com.squareup.moshi.Json;
+
+public class WebhookPayload {
+
+    @Json(name = "payload") private final WebhookEvent payload;
+
+    protected WebhookPayload(WebhookEvent payload) {
+        this.payload = payload;
+    }
+
+    public WebhookEvent getPayload() {
+        return payload;
+    }
+}
