+++ onfido-java/src/test/java/com/onfido/api/WebhookManagerTest.java
@@ -1,8 +1,8 @@
-package com.onfido.integration;
+package com.onfido.api;
 
 import com.onfido.JsonObject;
 import com.onfido.Onfido;
-import com.onfido.models.Applicant;
+import com.onfido.models.Webhook;
 import okhttp3.mockwebserver.MockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
 import org.testng.annotations.Test;
@@ -12,12 +12,12 @@ import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 
-public class ApplicantManagerTest extends ApiIntegrationTest {
+public class WebhookManagerTest extends ApiIntegrationTest {
 
     @Test
-    public void createApplicant() throws Exception {
+    public void createWebhook() throws Exception {
         String response = new JsonObject()
-                .add("first_name", "First")
+                .add("url", "url")
                 .toJson();
 
         MockWebServer server = mockRequestResponse(response);
@@ -27,26 +27,25 @@ public class ApplicantManagerTest extends ApiIntegrationTest {
                 .unknownApiUrl(server.url("/").toString())
                 .build();
 
-        Applicant applicant = onfido.applicant.create(Applicant.request()
-                .firstName("First"));
+        Webhook webhook = onfido.webhook.create(Webhook.request().url("url"));
 
         // Correct path
         RecordedRequest request = server.takeRequest();
-        assertEquals("/applicants/", request.getPath());
+        assertEquals("/webhooks/", request.getPath());
 
         // Correct request body
         String json = request.getBody().readUtf8();
         JsonObject jsonObject = JsonObject.parse(json);
-        assertEquals("First", jsonObject.get("first_name"));
+        assertEquals("url", jsonObject.get("url"));
 
         // Correct response body
-        assertEquals("First", applicant.getFirstName());
+        assertEquals("url", webhook.getUrl());
     }
 
     @Test
-    public void findApplicant() throws Exception {
+    public void findWebhook() throws Exception {
         String response = new JsonObject()
-                .add("first_name", "First")
+                .add("url", "url")
                 .toJson();
 
         MockWebServer server = mockRequestResponse(response);
@@ -56,20 +55,20 @@ public class ApplicantManagerTest extends ApiIntegrationTest {
                 .unknownApiUrl(server.url("/").toString())
                 .build();
 
-        Applicant applicant = onfido.applicant.find("id");
+        Webhook webhook = onfido.webhook.find("id");
 
         // Correct path
         RecordedRequest request = server.takeRequest();
-        assertEquals("/applicants/id", request.getPath());
+        assertEquals("/webhooks/id", request.getPath());
 
         // Correct response body
-        assertEquals("First", applicant.getFirstName());
+        assertEquals("url", webhook.getUrl());
     }
 
     @Test
-    public void updateApplicant() throws Exception {
+    public void updateWebhook() throws Exception {
         String response = new JsonObject()
-                .add("first_name", "First")
+                .add("url", "url")
                 .toJson();
 
         MockWebServer server = mockRequestResponse(response);
@@ -79,25 +78,24 @@ public class ApplicantManagerTest extends ApiIntegrationTest {
                 .unknownApiUrl(server.url("/").toString())
                 .build();
 
-        Applicant applicant = onfido.applicant.update("id", Applicant.request()
-                .firstName("First"));
+        Webhook webhook = onfido.webhook.update("id", Webhook.request().url("url"));
 
         // Correct path
         RecordedRequest request = server.takeRequest();
-        assertEquals("/applicants/id", request.getPath());
+        assertEquals("/webhooks/id", request.getPath());
 
         // Correct request body
         String json = request.getBody().readUtf8();
         JsonObject jsonObject = JsonObject.parse(json);
-        assertEquals("First", jsonObject.get("first_name"));
+        assertEquals("url", jsonObject.get("url"));
 
         // Correct response body
-        assertEquals("First", applicant.getFirstName());
+        assertEquals("url", webhook.getUrl());
     }
 
 
     @Test
-    public void deleteApplicant() throws Exception {
+    public void deleteWebhook() throws Exception {
         MockWebServer server = mockRequestResponse("");
 
         Onfido onfido = Onfido.builder()
@@ -105,34 +103,18 @@ public class ApplicantManagerTest extends ApiIntegrationTest {
                 .unknownApiUrl(server.url("/").toString())
                 .build();
 
-        onfido.applicant.delete("id");
+        onfido.webhook.delete("id");
 
         // Correct path
         RecordedRequest request = server.takeRequest();
-        assertEquals("/applicants/id", request.getPath());
+        assertEquals("/webhooks/id", request.getPath());
     }
 
     @Test
-    public void restoreApplicant() throws Exception {
-        MockWebServer server = mockRequestResponse("");
-
-        Onfido onfido = Onfido.builder()
-                .apiToken("token")
-                .unknownApiUrl(server.url("/").toString())
-                .build();
-
-        onfido.applicant.restore("id");
-
-        // Correct path
-        RecordedRequest request = server.takeRequest();
-        assertEquals("/applicants/id/restore", request.getPath());
-    }
-
-    @Test
-    public void listApplicants() throws Exception {
-        String response = new JsonObject().add("applicants", Arrays.asList(
-                new JsonObject().add("first_name", "First1").map,
-                new JsonObject().add("first_name", "First2").map))
+    public void listWebhooks() throws Exception {
+        String response = new JsonObject().add("webhooks", Arrays.asList(
+                new JsonObject().add("url", "url1").map,
+                new JsonObject().add("url", "url2").map))
                 .toJson();
 
         MockWebServer server = mockRequestResponse(response);
@@ -142,14 +124,14 @@ public class ApplicantManagerTest extends ApiIntegrationTest {
                 .unknownApiUrl(server.url("/").toString())
                 .build();
 
-        List<Applicant> applicants = onfido.applicant.list(1, 20, true);
+        List<Webhook> webhooks = onfido.webhook.list();
 
         // Correct path
         RecordedRequest request = server.takeRequest();
-        assertEquals("/applicants/?page=1&per_page=20&include_deleted=true", request.getPath());
+        assertEquals("/webhooks/", request.getPath());
 
         // Correct response body
-        assertEquals("First1", applicants.get(0).getFirstName());
-        assertEquals("First2", applicants.get(1).getFirstName());
+        assertEquals("url1", webhooks.get(0).getUrl());
+        assertEquals("url2", webhooks.get(1).getUrl());
     }
 }
